feat: Update TranslationController and related components to support decimal numbers for pluralization
- Modified the `number` parameter in `GetEntry` and `GetEntryUsingClient` methods to accept decimal values, enhancing pluralization support for fractional numbers.
- Updated XML documentation to clarify that the `number` parameter now supports both integer and decimal values.
- Adjusted related interfaces and classes across the SDK to ensure consistent handling of decimal numbers for translation entries.
- Added unit tests to validate the new functionality and ensure correct behavior with decimal inputs.

+++ src/Translaas.Client/ITranslaasClient.cs
@@ -16,7 +16,7 @@ public interface ITranslaasClient
     /// <param name="group">The translation group name.</param>
     /// <param name="entry">The translation entry key.</param>
     /// <param name="lang">The language code (e.g., "en", "fr").</param>
-    /// <param name="number">Optional number for pluralization.</param>
+    /// <param name="number">Optional number for pluralization. Supports both integer and decimal/fractional numbers (e.g., 1.31).</param>
     /// <param name="cancellationToken">Cancellation token.</param>
     /// <returns>The translated text.</returns>
     /// <exception cref="Translaas.Models.Errors.TranslaasApiException">Thrown when the API returns an error.</exception>
@@ -24,7 +24,7 @@ public interface ITranslaasClient
         string group,
         string entry,
         string lang,
-        int? number = null,
+        decimal? number = null,
         CancellationToken cancellationToken = default);
 
     /// <summary>
